The provided commit message does not match the diffs.
Based on the diffs, a more fitting message would be:

"Change method of fetching phone code in bsPhone"

The method of fetching phone code in the 'click' event function within jquery.bsPhone.js has been changed. Instead of extracting the text from the clicked dropdown item and trimming it, the entire HTML content of the dropdown item is now directly used.