Add active class to default country in bsPhone
The implementation of the phone code dropdown in the jquery.bsPhone.js file has been modified. Now, the 'active' CSS class is added to the dropdown item corresponding to the default country. This improves visibility and user experience, as the default country is now more distinctly marked in the dropdown menu.