Add active highlighting to default country in dropdown
The changes include an adjustment to the dropdown menu in jquery.bsPhone.js. The 'active' CSS class is now applied to the default country, enhancing its visibility within the dropdown menu and ultimately improving the user interface experience. The default country is now more distinguishable, allowing users to identify this choice quickly.