Remove 'ad.svg' flag icon
all flag icon files has been removed. This change might be a part of a larger refactoring, or it could be due to copyright issues or the flag being outdated or incorrect. The commit does not impact the functionality of the code base, but may affect the display of certain countries' flags in the application